+++ packages/adminui-vue/components/sidebar/USidebarItem.vue
@@ -7,6 +7,10 @@
     @contextmenu.native="contextDisabled ? null : contextOnTitle($event)"
   >
     <template slot="title">
+      <div
+        :style="{marginLeft}"
+        class="ub-sidebar__item-title__negative-margin"
+      />
       <i :class="getIcon(item)" />
       <span>{{ item.caption }}</span>
     </template>
@@ -20,6 +24,7 @@
       :item="child"
       :context-show="contextShow"
       :context-disabled="contextDisabled"
+      :level="level+1"
     />
   </el-submenu>
 
@@ -29,6 +34,10 @@
     @click="openLink(item)"
     @contextmenu.native="contextDisabled ? null : contextShow($event, item)"
   >
+    <div
+      :style="{marginLeft}"
+      class="ub-sidebar__item-title__negative-margin"
+    />
     <i :class="getIcon(item)" />
     <span slot="title">{{ item.caption }}</span>
   </el-menu-item>
@@ -38,9 +47,19 @@
 export default {
   name: 'USidebarItem',
   props: {
+    level: {
+      type: Number,
+      default: 1
+    },
     item: Object,
     contextShow: Function,
     contextDisabled: Boolean
+  },
+
+  computed: {
+    marginLeft () {
+      return -this.level * 10 + 'px'
+    }
   },
 
   methods: {
@@ -48,7 +67,7 @@ export default {
       $App.runShortcutCommand(ID, inWindow)
     },
 
-    getIcon(item) {
+    getIcon (item) {
       return item.iconCls || (item.parentID ? '' : (item.isFolder ? 'fa fa-folder' : 'fa fa-file-text-o'))
     },
 
@@ -69,15 +88,14 @@ export default {
   display: none;
   font-size: 14px;
   font-weight: 700;
+  padding: 8px;
+  height: auto;
   color: white;
-  padding: 0 20px;
-  height: 40px;
-  align-items: center;
   border-bottom: 1px solid rgba(255, 255, 255, 0.05);
 }
 
 .ub-sidebar__popup-menu .ub-submenu__on-colapsed__title{
-  display: flex;
+  display: block;
 }
 
 .ub-sidebar__popup-menu .el-menu-item ,
@@ -113,4 +131,10 @@ export default {
   display: flex;
   align-items: center;
 }
+
+.ub-sidebar .el-menu--collapse .ub-sidebar__item-title__negative-margin,
+.ub-sidebar__popup-menu .ub-sidebar__item-title__negative-margin{
+  display: none;
+}
+
 </style>